+++ providers/image.rb
@@ -1,16 +1,14 @@
-include Helpers::Docker
+include Docker::Helpers
 
 def load_current_resource
   wait_until_ready!
-  @current_resource = Chef::Resource::DockerImage.new(new_resource)
-  dimages = docker_cmd('images -a -notrunc')
+  @current_resource = Chef::Resource::DockerImage.new(new_resource.name)
+  dimages = docker_cmd('images -a --no-trunc')
   if dimages.stdout.include?(new_resource.image_name)
     dimages.stdout.each_line do |di_line|
       image = di(di_line)
-      unless image_id_matches?(image['id'])
-        next unless image_name_matches?(image['repository'])
-        next unless image_tag_matches_if_exists?(image['tag'])
-      end
+      next unless image_name_matches?(image['repository'])
+      next unless image_tag_matches_if_exists?(image['tag'])
       Chef::Log.debug('Matched docker image: ' + di_line.squeeze(' '))
       @current_resource.created(image['created'])
       @current_resource.repository(image['repository'])
@@ -66,9 +64,9 @@ action :pull_if_missing do
 end
 
 action :pull do
-  old_hash = docker_inspect_id(new_resource.image_name)
+  old_hash = docker_inspect_id(image_and_tag_arg)
   pull
-  new_hash = docker_inspect_id(new_resource.image_name)
+  new_hash = docker_inspect_id(image_and_tag_arg)
   new_resource.updated_by_last_action(new_hash != old_hash)
 end
 
@@ -140,7 +138,7 @@ def di(di_line)
   image
 end
 
-def command_timeout_error_message
+def command_timeout_error_message(cmd)
   <<-EOM
 
 Command timed out:
@@ -157,10 +155,12 @@ def image_and_tag_arg
 end
 
 def image_id_matches?(id)
+  return false unless id && new_resource.id
   id.start_with?(new_resource.id)
 end
 
 def image_name_matches?(name)
+  return false unless name && new_resource.image_name
   name.include?(new_resource.image_name)
 end