+++ providers/registry.rb
@@ -1,7 +1,7 @@
-include Helpers::Docker
+include Docker::Helpers
 
 def load_current_resource
-  @current_resource = Chef::Resource::DockerRegistry.new(new_resource)
+  @current_resource = Chef::Resource::DockerRegistry.new(new_resource.name)
   wait_until_ready!
   dockercfg = dockercfg_parse
   if dockercfg && login_matches(dockercfg[new_resource.server])
@@ -20,7 +20,7 @@ action :login do
   end
 end
 
-def command_timeout_error_message
+def command_timeout_error_message(cmd)
   <<-EOM
 
 Command timed out: