+++ recipes/runit.rb
@@ -1,9 +1,9 @@
 # Package installation automatically starts docker service
 # So let's stop and disable it.
 service 'lxc-docker' do
-  pattern Helpers::Docker.executable(node)
+  pattern Docker::Helpers.executable(node)
   provider Chef::Provider::Service::Upstart if node['platform'] == 'ubuntu'
-  service_name Helpers::Docker.docker_service(node)
+  service_name Docker::Helpers.docker_service(node)
   action [:stop, :disable]
 end
 
@@ -12,6 +12,6 @@ include_recipe 'runit'
 runit_service 'docker' do
   default_logger true
   options(
-    'daemon_options' => Helpers::Docker.daemon_cli_args(node)
+    'daemon_options' => Docker::Helpers.daemon_cli_args(node)
   )
 end