+++ spec/binary_spec.rb
@@ -1,19 +1,25 @@
 require 'spec_helper'
 
 describe 'docker::binary' do
-  let(:chef_run) do
-    ChefSpec::Runner.new.converge(described_recipe)
-  end
 
-  it 'downloads docker binary' do
-    expect(chef_run).to create_remote_file_if_missing('/usr/bin/docker')
+  context 'by default' do
+    let(:chef_run) do
+      ChefSpec::Runner.new(platform: 'ubuntu', version: '12.04') do |node|
+        node.automatic['kernel']['release'] = '3.8.0'
+      end.converge(described_recipe)
+    end
+
+    it 'downloads docker binary' do
+      expect(chef_run).to create_remote_file_if_missing('/usr/bin/docker')
+    end
   end
 
   context 'when install_dir is set' do
     let(:chef_run) do
-      runner = ChefSpec::Runner.new
-      runner.node.set['docker']['install_dir'] = '/tmp'
-      runner.converge(described_recipe)
+      ChefSpec::Runner.new(platform: 'ubuntu', version: '12.04') do |node|
+        node.normal['docker']['install_dir'] = '/tmp'
+        node.automatic['kernel']['release'] = '3.8.0'
+      end.converge(described_recipe)
     end
 
     it 'downloads docker binary to install_dir' do
@@ -23,9 +29,10 @@ describe 'docker::binary' do
 
   context 'when install_type is binary' do
     let(:chef_run) do
-      runner = ChefSpec::Runner.new
-      runner.node.set['docker']['install_type'] = 'binary'
-      runner.converge(described_recipe)
+      ChefSpec::Runner.new(platform: 'ubuntu', version: '12.04') do |node|
+        node.normal['docker']['install_type'] = 'binary'
+        node.automatic['kernel']['release'] = '3.8.0'
+      end.converge(described_recipe)
     end
 
     it 'downloads docker binary to install_dir' do