+++ spec/default_spec.rb
@@ -15,32 +15,68 @@ describe 'docker::default' do
     allow(uname).to receive(:stdout).and_return('3.')
 
     stub_command('modprobe -n -v aufs').and_return('')
+    stub_command('mountpoint -q /sys/fs/cgroup').and_return('')
 
     # TODO: Contribute back to golang cookbook
-    stub_command('/usr/local/go/bin/go version | grep "go1.2 "').and_return('1.2')
+    stub_command('/usr/local/go/bin/go version | grep "go1.2.2 "').and_return('1.2.2')
   end
 
-  let(:chef_run) do
-    ChefSpec::Runner.new.converge(described_recipe)
+  context 'when running on ubuntu 12.04' do
+    let(:chef_run) do
+      ChefSpec::Runner.new(platform: 'ubuntu', version: '12.04') do |node|
+        node.set['docker']['alert_on_error_action'] = :warn
+      end.converge(described_recipe)
+    end
+
+    it 'should make sure kernel >= 3.8'
+
+    it 'includes the apt recipe' do
+      expect(chef_run).to include_recipe('apt')
+    end
+
+    it 'installs the apt-transport-https package' do
+      expect(chef_run).to install_package('apt-transport-https')
+    end
+
+    it 'installs the bsdtar package' do
+      expect(chef_run).to install_package('bsdtar')
+    end
   end
 
-  it 'includes the apt recipe' do
-    expect(chef_run).to include_recipe('apt')
-  end
+  context 'when running on debian 7.4' do
+    let(:chef_run) do
+      ChefSpec::Runner.new(platform: 'debian', version: '7.4') do |node|
+        node.set['docker']['alert_on_error_action'] = :warn
+      end.converge(described_recipe)
+    end
 
-  it 'installs the apt-transport-https package' do
-    expect(chef_run).to install_package('apt-transport-https')
-  end
+    it 'includes the apt recipe' do
+      expect(chef_run).to include_recipe('apt')
+    end
 
-  it 'installs the bsdtar package' do
-    expect(chef_run).to install_package('bsdtar')
+    it 'installs the apt-transport-https package' do
+      expect(chef_run).to install_package('apt-transport-https')
+    end
+
+    it 'installs the bsdtar package' do
+      expect(chef_run).to install_package('bsdtar')
+    end
+
+    it 'sets net.ipv4.ip_forward to 1' do
+      expect(chef_run).to apply_sysctl_param('net.ipv4.ip_forward').with(value: 1)
+    end
+
+    it 'sets net.ipv6.conf.all.forwarding to 1' do
+      expect(chef_run).to apply_sysctl_param('net.ipv6.conf.all.forwarding').with(value: 1)
+    end
   end
 
   context 'when exec_driver is lxc' do
     let(:chef_run) do
-      runner = ChefSpec::Runner.new
-      runner.node.set['docker']['exec_driver'] = 'lxc'
-      runner.converge(described_recipe)
+      ChefSpec::Runner.new do |node|
+        node.set['docker']['alert_on_error_action'] = :warn
+        node.set['docker']['exec_driver'] = 'lxc'
+      end.converge(described_recipe)
     end
 
     it 'includes the docker::cgroups recipe' do
@@ -54,9 +90,38 @@ describe 'docker::default' do
 
   context 'when install_type is binary' do
     let(:chef_run) do
-      runner = ChefSpec::Runner.new
-      runner.node.set['docker']['install_type'] = 'binary'
-      runner.converge(described_recipe)
+      ChefSpec::Runner.new do |node|
+        node.set['docker']['alert_on_error_action'] = :warn
+        node.set['docker']['install_type'] = 'binary'
+      end.converge(described_recipe)
+    end
+
+    it 'includes iptables cookbook' do
+      expect(chef_run).to include_recipe('iptables')
+    end
+
+    it 'includes git cookbook' do
+      expect(chef_run).to include_recipe('git')
+    end
+
+    it 'installs procps package' do
+      expect(chef_run).to install_package('procps')
+    end
+
+    it 'installs the XZ Utilities package' do
+      expect(chef_run).to install_package('xz-utils')
+
+      rhel_run = ChefSpec::Runner.new(platform: 'redhat', version: '6.5') do |node|
+        node.set['docker']['alert_on_error_action'] = :warn
+        node.set['docker']['install_type'] = 'binary'
+        node.automatic['kernel']['release'] = '3.8.0'
+      end.converge(described_recipe)
+
+      expect(rhel_run).to install_package('xz')
+    end
+
+    it 'create cgroupfs-mount script' do
+      expect(chef_run).to render_file('/usr/local/bin/cgroupfs-mount')
     end
 
     it 'includes the docker::binary recipe' do
@@ -64,24 +129,42 @@ describe 'docker::default' do
     end
   end
 
-  context 'when install_type is binary and storage_driver is aufs' do
+  context 'when running on ubuntu < 13.10' do
+    context 'and install_type is binary and storage_driver is aufs' do
+      let(:chef_run) do
+        ChefSpec::Runner.new do |node|
+          node.set['docker']['alert_on_error_action'] = :warn
+          node.set['docker']['install_type'] = 'binary'
+          node.set['docker']['storage_driver'] = 'aufs'
+        end.converge(described_recipe)
+      end
+
+      it 'includes the docker::aufs recipe' do
+        expect(chef_run).to include_recipe('docker::aufs')
+      end
+    end
+  end
+
+  context 'when running on ubuntu > 13.10' do
     let(:chef_run) do
-      runner = ChefSpec::Runner.new
-      runner.node.set['docker']['install_type'] = 'binary'
-      runner.node.set['docker']['storage_driver'] = 'aufs'
-      runner.converge(described_recipe)
+      ChefSpec::Runner.new(platform: 'ubuntu', version: '14.04') do |node|
+        node.set['docker']['alert_on_error_action'] = :warn
+        node.set['docker']['install_type'] = 'binary'
+        node.set['docker']['storage_driver'] = 'aufs'
+      end.converge(described_recipe)
     end
 
-    it 'includes the docker::aufs recipe' do
-      expect(chef_run).to include_recipe('docker::aufs')
+    it 'should not manage the storagedriver' do
+      expect(chef_run).not_to include_recipe('docker::aufs')
     end
   end
 
   context 'when install_type is source' do
     let(:chef_run) do
-      runner = ChefSpec::Runner.new
-      runner.node.set['docker']['install_type'] = 'source'
-      runner.converge(described_recipe)
+      ChefSpec::Runner.new do |node|
+        node.set['docker']['alert_on_error_action'] = :warn
+        node.set['docker']['install_type'] = 'source'
+      end.converge(described_recipe)
     end
 
     it 'includes the git recipe' do
@@ -97,59 +180,30 @@ describe 'docker::default' do
     end
   end
 
-  it 'includes the docker::package recipe' do
-    expect(chef_run).to include_recipe('docker::package')
-  end
-
-  context 'when group_members is not empty' do
+  context 'when the install_type is package' do
     let(:chef_run) do
-      runner = ChefSpec::Runner.new
-      runner.node.set['docker']['group_members'] = ['vagrant']
-      runner.converge(described_recipe)
+      ChefSpec::Runner.new do |node|
+        node.set['docker']['alert_on_error_action'] = :warn
+      end.converge(described_recipe)
     end
 
-    it 'includes the docker::group recipe' do
-      expect(chef_run).to include_recipe('docker::group')
-    end
-  end
-
-  it 'includes the docker::upstart recipe' do
-    expect(chef_run).to include_recipe('docker::upstart')
-  end
-
-  context 'when init_type is runit' do
-    let(:chef_run) do
-      runner = ChefSpec::Runner.new
-      runner.node.set['docker']['init_type'] = 'runit'
-      runner.converge(described_recipe)
-    end
-
-    it 'includes the docker::runit recipe' do
-      expect(chef_run).to include_recipe('docker::runit')
+    it 'includes the docker::package recipe' do
+      expect(chef_run).to include_recipe('docker::package')
     end
   end
 
-  context 'when init_type is systemd' do
-    let(:chef_run) do
-      runner = ChefSpec::Runner.new
-      runner.node.set['docker']['init_type'] = 'systemd'
-      runner.converge(described_recipe)
-    end
+  %w( runit systemd sysv ).each do |init|
+    context "when init_type is #{init}" do
+      let(:chef_run) do
+        ChefSpec::Runner.new do |node|
+          node.set['docker']['alert_on_error_action'] = :warn
+          node.set['docker']['init_type'] = init
+        end.converge(described_recipe)
+      end
 
-    it 'includes the docker::systemd recipe' do
-      expect(chef_run).to include_recipe('docker::systemd')
-    end
-  end
-
-  context 'when init_type is sysv' do
-    let(:chef_run) do
-      runner = ChefSpec::Runner.new
-      runner.node.set['docker']['init_type'] = 'sysv'
-      runner.converge(described_recipe)
-    end
-
-    it 'includes the docker::sysv recipe' do
-      expect(chef_run).to include_recipe('docker::sysv')
+      it "includes the docker::#{init} recipe" do
+        expect(chef_run).to include_recipe("docker::#{init}")
+      end
     end
   end
 end