+++ spec/upstart_spec.rb
@@ -8,16 +8,28 @@ describe 'docker::upstart' do
   it 'creates the docker Upstart template' do
     expect(chef_run).to create_template('/etc/init/docker.conf')
     expect(chef_run).to render_file('/etc/init/docker.conf').with_content(
-      /"\$UPSTART_JOB" -d \$DOCKER_OPTS/)
+      /exec "\$DOCKER" -d \$DOCKER_OPTS/)
   end
 
-  it 'creates the docker sysconfig template' do
-    expect(chef_run).to create_template('/etc/default/docker')
-    expect(chef_run).to render_file('/etc/default/docker').with_content(
-      %r{^DOCKER="/usr/bin/docker"$})
-    resource = chef_run.template('/etc/default/docker')
-    expect(resource).to notify('service[docker]').to(:stop).immediately
-    expect(resource).to notify('service[docker]').to(:start).immediately
+  context 'when running on debian/ubuntu' do
+    it 'creates the docker sysconfig template in /etc/default' do
+      expect(chef_run).to create_template('/etc/default/docker')
+      expect(chef_run).to render_file('/etc/default/docker').with_content(
+        %r{^DOCKER="/usr/bin/docker"$})
+      resource = chef_run.template('/etc/default/docker')
+      expect(resource).to notify('service[docker]').to(:stop).immediately
+      expect(resource).to notify('service[docker]').to(:start).immediately
+    end
+  end
+
+  context 'when running on non debian/ubuntu' do
+    let(:chef_run) do
+      ChefSpec::Runner.new(platform: 'centos', version: '6.5').converge(described_recipe)
+    end
+
+    it 'creates the docker sysconfig template in /etc/sysconfig' do
+      expect(chef_run).to create_template('/etc/sysconfig/docker')
+    end
   end
 
   context 'when api_enable_cors is set' do